+++ tests/set_creators/set_creator_functions_test.py
@@ -1,4 +1,6 @@
 import unittest
+
+import numpy as np
 from dexterous_bioprosthesis_2021_raw_datasets.set_creators.np_signal_extractors.np_signal_extractor_ar import NpSignalExtractorAr
 from dexterous_bioprosthesis_2021_raw_datasets.set_creators.np_signal_extractors.np_signal_extractor_mav import NpSignalExtractorMav
 from dexterous_bioprosthesis_2021_raw_datasets.set_creators.np_signal_extractors.np_signal_extractor_ssc import NpSignalExtractorSsc
@@ -22,5 +24,5 @@ class SetCreatorFunctionsTest(SetCreatorTest):
             ]
         return extractors
     
-    def generate_sample_data(self):
-        return generate_sample_data(samples_number=1000)
+    def generate_sample_data(self, dtype=np.double):
+        return generate_sample_data(samples_number=1000, dtype=dtype)
